feat: 태그 입력 base 작업
