+++ frontend/src/views/components/post/Post.css
@@ -67,12 +67,18 @@
   font-weight: bold;
 }
 
-.contentLink {
+.contentLink,
+.contentDisable {
   position: absolute;
+  top: 0px;
   left: 0px;
   width: 100%;
   height: 100%;
   z-index: 1;
+}
+
+.contentDisable {
+  z-index: 9999;
 }
 
 .contentTitle {
