+++ frontend/src/views/pages/postPages/PostPage.css
@@ -39,19 +39,6 @@
 }
 
 /* WritePage  */
-.postSubmitBtn {
-  color: #fdfeff;
-  font-weight: bold;
-  font-size: 15px;
-  position: relative;
-  top: 35px;
-  width: 250px;
-  height: 35px;
-  border-radius: 15px;
-  background: #1d9acb;
-  box-shadow: 0px 0px 20px 0px rgba(80, 80, 80, 20%);
-}
-
 .writeHead {
   position: fixed;
   top: 22px;
@@ -64,48 +51,4 @@
 
 .emptyTitle {
   color: #e6e8ea;
-}
-
-.postInput {
-  box-sizing: border-box;
-  border: 0;
-  border-radius: 12.5px;
-  width: 250px;
-  height: 30px;
-  margin-top: 10px;
-  background: #f0f3f6;
-  padding: 0 10px 0 10px;
-  text-align: center;
-}
-
-.typeSelector {
-  margin-top: 10px;
-}
-
-.typeBtn {
-  color: #1d9acb;
-  background: #fdfeff;
-  border: 1px solid #1d9acb !important;
-  font-weight: bold;
-  width: 84px;
-  height: 30px;
-}
-
-.selectedType {
-  color: #fdfeff;
-  background: #1d9acb;
-}
-
-.leftT {
-  border-top-left-radius: 12.5px;
-  border-bottom-left-radius: 12.5px;
-}
-
-.centerT {
-  margin: 0 -1px 0 -1px;
-}
-
-.rightT {
-  border-top-right-radius: 12.5px;
-  border-bottom-right-radius: 12.5px;
 }
